Rename backend packages
backend/git -> backend/sync/git
backend/gpg -> backend/crypto/gpg
