+++ backend/crypto/gpg/cli/version.go
