+++ utils/jsonapi/api_test.go
@@ -12,7 +12,7 @@ import (
 	"strings"
 	"testing"
 
-	gpgmock "github.com/justwatchcom/gopass/backend/gpg/mock"
+	gpgmock "github.com/justwatchcom/gopass/backend/crypto/gpg/mock"
 	"github.com/justwatchcom/gopass/config"
 	"github.com/justwatchcom/gopass/store/root"
 	"github.com/justwatchcom/gopass/store/secret"
